// Client/src/app/layout/profile/profile.interface.ts
/**
 * Created by Moayeri on 05/27/2018.
 */
export interface IProfile {
    userId: number;
    displayName: string;
    realName: string;
    birthday: string;
    height: number;
    gender: number;
    ethnicity: number;
    religion: number;
    figure: number;
    maritalStatus: number;
    occupation: string;
    aboutMe: string;
    location: number;
}